requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.6
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
